refactor(web): share the grid card's draft handlers
Dropping the eventType argument left MainGridEvents and AllDayEvents with
byte-identical copies of the lookup, both handlers, and the four-line
comment explaining why they must not use `start`. One hook owns that shape
now, so the rule lives in one place instead of two that can drift.

useCallback: three in the hook, same as the copies they replace - the
handlers are props on memoized cards whose comparators would otherwise
re-render every card on each parent render.
